refactor: move crypto functionality to folder
