+++ .gitignore
@@ -1,2 +1,3 @@
 .idea
 target
+.DS_Store
