+++ cli/src/file.rs
@@ -3,7 +3,7 @@ use std::error::Error;
 use std::fs::File;
 use std::io::{Read, Write};
 use app_dirs2::{app_dir, AppDataType};
-use common::EncryptedValue;
+use common::crypto::asymmetric::EncryptedValue;
 use crate::{APP_INFO, CredentialsMap};
 
 const DATA_DIR: &str = "data";
